remove aliases
